Moving ErrorHandlingSelection to ErrorHandling folder
